chore: update README.md and GIM defaul url
Updated defaul URL for GIM because the old one will stop working
soon. The new url is https://gim.stone.com.br/api. The package
version was also updated and the readme was update in order to
show the getUser method of the management class and the way of
instantiating the library.

Resolves: #5